### 2024-09-22 #658 @alex-koiter
mbquartR: Finding Manitoba Quarter Sections
Submitting Author Name: Alexander Koiter
Submitting Author Github Handle: @alex-koiter
Other Package Authors Github handles: (comma separated, delete if none)
Repository:  https://github.com/alex-koiter/mbquartR
Version submitted:  0.0.0.9000
Submission type: Standard
Version accepted: TBD

---

Package: mbquartR
Title: Finding Manitoba Quarter Sections
Version: 0.0.0.9000
    person("Alex", "Koiter", , "koitera@brandonu.ca", role = c("aut", "cre"),
Description: This package has three main functions: 1) find the coordinates of a quarter sections given the legal land description (e.g., "NE-11-33-29W"); 2) find the legal land description using coordinates (lat and long); and 3) plot these points on a map.
    curl (>= 5.2.1),
    purrr (>= 1.0.2),
    sf (>= 1.0.16),
    tibble (>= 3.2.1),
    tidyr (>= 1.3.1),
    stringr (>= 1.5.1),
    units (>= 0.8.5),
    mapview (>= 2.11.2),
    readr (>= 2.1.4),
    crayon (>= 1.5.2),
    rlang

This package enables users to download the Manitoba Original Survey Legal Descriptions data from the Manitoba government geoportal. It includes functions that allow users to easily search for parcels of land (i.e., quarter sections) by either the legal land description or by coordinates and plot them on an interactive map.

The target audience are those who work with geospatial data in Manitoba, particularly those who are working with rural or farm parcels of land where the legal land description is commonly used as the method of identifying the location. Also anyone who needs to quickly go back and forth between geographic coordinates and the legal land description. Currently, there are a few web-based tools (https://geoportal.gov.mb.ca/datasets/11fa11f9015b45438d6493dcb3d8071c/, https://www.masc.mb.ca/masc.nsf/land_parcel_info.html, https://legallandconverter.com/p4.html) that allow users to locate land parcels but do not allow for batch processing, do not return coordinates, and/or are fee based. Being able to convert back and forth is also helpful for scientist preparing outreach materials.

No

NA. This package provides access to publicly available data.

https://github.com/ropensci/software-review/issues/649

`pkgcheck` passes fully with github actions but fails locally because it doesn't think I have CI (but I do).
