### 2024-09-22 #658 @adamhsparks
Hi @alex-koiter, the rOpenSci bot says this is all good, but when I run checks locally, I get an error.

```r
> devtools::check()
══ Documenting ════════════════════════════════════════════════════════════════════════════
ℹ Installed roxygen2 version (7.3.2) doesn't match required (7.3.1)
✖ `check()` will not re-document this package
══ Building ═══════════════════════════════════════════════════════════════════════════════
Setting env vars:
• CFLAGS    : -Wall -pedantic -fdiagnostics-color=always
• CXXFLAGS  : -Wall -pedantic -fdiagnostics-color=always
• CXX11FLAGS: -Wall -pedantic -fdiagnostics-color=always
• CXX14FLAGS: -Wall -pedantic -fdiagnostics-color=always
• CXX17FLAGS: -Wall -pedantic -fdiagnostics-color=always
• CXX20FLAGS: -Wall -pedantic -fdiagnostics-color=always
── R CMD build ────────────────────────────────────────────────────────────────────────────
✔  checking for file ‘/Users/adamsparks/tmp/mbquartR/DESCRIPTION’ ...
─  preparing ‘mbquartR’:
E  creating vignettes (1.5s)
   --- re-building ‘mbquartR.Rmd’ using rmarkdown

   Quitting from lines 69-71 [example] (mbquartR.Rmd)
   Error: processing vignette 'mbquartR.Rmd' failed with diagnostics:
   Data doesn't exist, please download with `quarters_dl()` first
   --- failed re-building ‘mbquartR.Rmd’

   SUMMARY: processing the following file failed:
     ‘mbquartR.Rmd’

   Error: Vignette re-building failed.
   Execution halted
Error:
! ! System command 'R' failed
Show Traceback
```

Please ignore the roxygen2 comment, I'm concerned about the Error creating vignettes.
