### 2024-09-23 #658 @alex-koiter
Thanks @adamhsparks! The issue was that the vignette needed to access data that hadn't been downloaded yet. I have now created some sample data that the vignette uses. Also found some additional typos and made a few other small changes. It should work now.
