### 2024-10-01 #658 @jhollist
@alex-koiter I am just now starting to look at this (apologies!!)  I still get the the same error as @adamhsparks with the vignette build.  I have cloned and attempted to build https://github.com/alex-koiter/mbquartR and get the error. Perhaps the fix wasn't pushed?

Also I see no documentation on your non-exported functions (Statistical Properties in https://github.com/ropensci/software-review/issues/658#issuecomment-2366968970).  While not technically required, it would be  an easy area of improvement to add documentation on these non-exported functions.
