### 2024-10-02 #658 @alex-koiter
@jhollist It turns out that I only partially resolved the issue (or fixed a different one). I have resolved the error so you should be able to clone and build the package now. I will take me a few days to get the documentation for the internal functions completed and I will let you know once that is done.
Thanks!
