### 2024-10-03 #658 @jhollist
And FYI, check runs now on my machine!  I did see this:

---
─  checking examples ... [43s] OK (43.4s)
   Examples with CPU (user + system) or elapsed time > 5s
                 user system elapsed
   map_quarter  23.53   2.08   28.48
   search_coord 10.39   0.65    8.47
---
I know this is system specific so might not be a real issue, but if there is an easy way to have the example run a bit faster that would be good.
