### 2024-10-18 #658 @alex-koiter
@jhollist The internal function documentation is now complete. As for the run time, I think this may be an artifact of how the functions are written. This is where the reviewers may be able to provide some suggestions.
Thanks!
