### 2024-12-10 #658 @EmilyMarkowitz-NOAA
---
title: "review"
  rmarkdown::md_document:
    pandoc_args: [
      "--wrap=none"
    ]
---

- [X] **Installation:** Installation succeeds as documented. _See output below._
- [ ] **Automated tests:** Unit tests cover essential functions of the package and a reasonable range of inputs and conditions. All tests pass on the local machine. _There are some issues with the tests, see output below._

_No need, but thanks for offering._
---

In addition to agreeing with the first reviewer's comments, I have a few to add, which I've mostly left in the numbered bullets below. In general, and similarly to the other reviewer, I was able to install the R package, run the functions as expected, thought the package worked well and was well documented. Though this R Package has a very specific audience, it sounds like it will be an easy and helpful too for them. I share similar sentiments that the README could include more details and context and had the same errors with the test checks that the other reviewer had.

Thanks for the opportunity to review and let me know if you have any questions! This is my first time reviewing an R package for ROpenSci, so please let me know if there is anything I missed or forgot to cover. Good luck, and hope this helps! Thanks for making a package to help the folks of Manitoba!

1. Readme could include more information about the institution, author ([example](https://github.com/afsc-gap-products/gap_products?tab=readme-ov-file#this-code-is-primarally-maintained-by)), additional resources, the government legal (example [1](https://github.com/afsc-gap-products/gap_products?tab=readme-ov-file#access-constraints), [2](https://github.com/afsc-gap-products/gap_products?tab=readme-ov-file#noaa-license)), and access constraints to this data ([example](https://github.com/afsc-gap-products/gap_products?tab=readme-ov-file#access-constraints)). I've included links here to a repository I'm working on at NOAA for reference.
2. The first reviewer made a great note about maintainer redundancy. Possibly related to or fixed by stating your affiliation (as I noted above), but sharing more about who you are and your/your affiliation's commitment to this package would help users know its priority for upkeep to maintainers and who might take over if you pass the package along to someone else to maintain.
3. It may be worthwhile to choose a licence like the GNU General Public Licence v3.0 or MIT Licence for your work, rather than just the two "YEAR" and "COPYWRITE HOLDER" lines. They're fine, but its my understanding that the ones I mentioned are standard practice and are recommended at my organization.
4. Cite the contributor guide (a `CONTRIBUTING.md`) on your README. It took me a minute to find in your pkgdown site. It may be worthwhile to also add some other types of contributor guidelines, like those from repos we use at [NOAA](https://github.com/nmfs-opensci/Contributor-Guide) (pruely for reference).
5. Possibly add a summary map of Manitoba with all of the different areas to it, so users can get a sense for the breadth of what the R Package covers. As a user unfamiliar with Manitoba's quarter sections system, I had to make some assumptions about what "quarters" meant and how they were used (is it the same thing as or related to [this](https://en.wikipedia.org/wiki/Dominion_Land_Survey)?).
6. Some of the bullets appear to be a bit messed up on this [page](https://alexkoiter.ca/mbquartR/reference/mbquartR-package.html). Also, I think the contents of this page could be added directly to a new or as a summary at the top of the [existing vingette](https://alexkoiter.ca/mbquartR/articles/mbquartR.html).
7. The [`quarters_dl()` example](https://alexkoiter.ca/mbquartR/reference/quarters_dl.html#arguments) seems to be a bit misprinted and I think would benefit from a . I think you want to use `\dontrun{}` like this:

```
#'@examples
#'\dontrun{
#'geocode("3817 Spruce St, Philadelphia, PA 19104")
#'geocode("Philadelphia, PA")
#'dat <- data.frame(value=runif(3),address=c("3817 Spruce St, Philadelphia, PA 19104","Philadelphia, PA","Neverneverland"))
#'geocode(dat)
#'}
```
8. In general, there are a few sentences throughout the documentation that don't end with periods (".").
9. I agree with the other reviewer's second comment. I'll add to that by noting: Specify the bounds and data type in the [`search_coords()` arguments examples](https://alexkoiter.ca/mbquartR/reference/search_coord.html#arguments). Perhaps something like `Numeric. Latitude (decimal degree) with resolution up to one hundred thousandth of a degree. Values outside of bounds will return error (49 to 59.5°N). ` I added the resolution and the "will return error" ideas in there, feel free to use or do something different. I used the general boundaries of manatoba here, for example.
10. In the `search_legal` function, I would add a link to where a user can look up the legal land descriptions.
11. Add more information and a citation for the [open data source](https://geoportal.gov.mb.ca) you are using. The only place I found it listed was in the [function file](https://github.com/alex-koiter/mbquartR/blob/6fd24396898b0c3e1f4ad2ffb772c3a29d3a2fca/R/download.R#L101-L102).
12. In addition to the other reviewer's 5th comment, it would be great to include a section that 1) identifies work that has used this package and 2) create a vingette of a specific example with storyline using this package (e.g., "A farmer wants to know ..."). Some of the content provided in the ROpenSci application could be helpful for this.
13. A package like this seems ripe for an R Shiny app - something that can be interactive for the non-coder. Just something to consider in a next stage of the package's development.
14. Seconding the first reviewer's comment 6). I am also not familiar with the legal land descriptions in Manitoba, but the examples provided in the vignette worked well for me without any errors.
15.  The first reviewer made a great point to ask how often the package data will be upated. It would be good to note this in the documentation for users so they can anticipate future changes.
16. See check run notes below. Specifically, I got two errors when I tried to run the package test:

**A. "search_coord() returns dataframe"**
`q` is a `tibble()`, not a dataframe
```
-- Error: search_coord() returns dataframe -------------------------------------
Error in `waldo::compare(x, y, ..., x_arg = x_arg, y_arg = y_arg)`: `tolerance` must be a number or `NULL`, not `TRUE`.
    x
 1. \-testthat::expect_equal(...)
 2.   \-testthat:::expect_waldo_equal("equal", act, exp, info, ..., tolerance = tolerance)
 3.     \-testthat:::waldo_compare(...)
 4.       \-waldo::compare(tolerance = NULL)
 5.         \-waldo:::check_number_decimal(tolerance, allow_null = TRUE, min = 0)
 6.           \-waldo:::.stop_not_number(...)
 7.             \-waldo:::stop_input_type(...)
 8.               \-rlang::abort(message, ..., call = call, arg = arg)

! Test failed
    x
 1. +-testthat::test_that(...)
 2. | \-withr (local) `<fn>`()
 3. \-reporter$stop_if_needed()
 4.   \-rlang::abort("Test failed", call = NULL)
```

**B. "search_legal() returns dataframe**
`q` is a `tibble()`, not a dataframe
```
-- Error: search_legal() returns dataframe -------------------------------------
Error in `waldo::compare(x, y, ..., x_arg = x_arg, y_arg = y_arg)`: `tolerance` must be a number or `NULL`, not `TRUE`.
    x
 1. \-testthat::expect_equal(...)
 2.   \-testthat:::expect_waldo_equal("equal", act, exp, info, ..., tolerance = tolerance)
 3.     \-testthat:::waldo_compare(...)
 4.       \-waldo::compare(tolerance = NULL)
 5.         \-waldo:::check_number_decimal(tolerance, allow_null = TRUE, min = 0)
 6.           \-waldo:::.stop_not_number(...)
 7.             \-waldo:::stop_input_type(...)
 8.               \-rlang::abort(message, ..., call = call, arg = arg)

! Test failed
    x
 1. +-testthat::test_that(...)
 2. | \-withr (local) `<fn>`()
 3. \-reporter$stop_if_needed()
 4.   \-rlang::abort("Test failed", call = NULL)
```

#### Check runs

<summary>R Session Info</summary>

`sessionInfo()`

```
R version 4.4.1 (2024-06-14 ucrt)
Platform: x86_64-w64-mingw32/x64
Running under: Windows 10 x64 (build 19045)

[1] LC_COLLATE=English_United States.utf8  LC_CTYPE=English_United States.utf8    LC_MONETARY=English_United States.utf8
[4] LC_NUMERIC=C                           LC_TIME=English_United States.utf8

time zone: America/Los_Angeles
tzcode source: internal

[1] mbquartR_0.0.0.9000

 [1] miniUI_0.1.1.1    compiler_4.4.1    promises_1.3.2    Rcpp_1.0.13-1     rcmdcheck_1.4.0   callr_3.7.6       later_1.4.1
 [8] fastmap_1.2.0     mime_0.12         R6_2.5.1          curl_6.0.1        htmlwidgets_1.6.4 tibble_3.2.1      desc_1.4.3
[15] profvis_0.4.0     rprojroot_2.0.4   shiny_1.9.1       pillar_1.9.0      rlang_1.1.4       utf8_1.2.4        cachem_1.1.0
[22] httpuv_1.6.15     fs_1.6.5          pkgload_1.4.0     memoise_2.0.1     cli_3.6.3         withr_3.0.2       magrittr_2.0.3
[29] ps_1.8.1          processx_3.8.4    digest_0.6.37     rstudioapi_0.17.1 xtable_1.8-4      remotes_2.5.0     devtools_2.4.5
[36] lifecycle_1.0.4   prettyunits_1.2.0 vctrs_0.6.5       glue_1.8.0        xopen_1.0.1       urlchecker_1.0.1  sessioninfo_1.2.2
[43] pkgbuild_1.4.5    fansi_1.0.6       purrr_1.0.2       tools_4.4.1       usethis_3.1.0     pkgconfig_2.0.3   ellipsis_0.3.2
[50] htmltools_0.5.8.1
```

<summary>Clean and install: success</summary>

```
==> Rcmd.exe INSTALL --preclean --no-multiarch --with-keep.source mbquartR

* installing to library 'C:/Users/emily.markowitz/AppData/Local/R/win-library/4.4'
* installing *source* package 'mbquartR' ...
** R
** data
*** moving datasets to lazyload DB
** byte-compile and prepare package for lazy loading
** help
*** installing help indices
** building package indices
** installing vignettes
** testing if installed package can be loaded from temporary location
** testing if installed package can be loaded from final location
** testing if installed package keeps a record of temporary installation path
* DONE (mbquartR)
```

<summary>Devtools output: 1 error (from test checks)</summary>

`devtools::check()`

```
══ Documenting ═══════════════════════════════════════════════════════════════════════════════════════════════════════════════════════════════════
ℹ Installed roxygen2 version (7.3.2) doesn't match required (7.3.1)
✖ `check()` will not re-document this package
══ Building ══════════════════════════════════════════════════════════════════════════════════════════════════════════════════════════════════════
Setting env vars:
• CFLAGS    : -Wall -pedantic -fdiagnostics-color=always
• CXXFLAGS  : -Wall -pedantic -fdiagnostics-color=always
• CXX11FLAGS: -Wall -pedantic -fdiagnostics-color=always
• CXX14FLAGS: -Wall -pedantic -fdiagnostics-color=always
• CXX17FLAGS: -Wall -pedantic -fdiagnostics-color=always
• CXX20FLAGS: -Wall -pedantic -fdiagnostics-color=always
── R CMD build ───────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────
✔  checking for file 'C:\Users\emily.markowitz\Work\projects\mbquartR/DESCRIPTION' (455ms)
─  preparing 'mbquartR':
✔  creating vignettes (19.9s)
─  checking for LF line-endings in source and make files and shell scripts (1.2s)
─  building 'mbquartR_0.0.0.9000.tar.gz'

══ Checking ══════════════════════════════════════════════════════════════════════════════════════════════════════════════════════════════════════
Setting env vars:
• _R_CHECK_CRAN_INCOMING_REMOTE_               : FALSE
• _R_CHECK_CRAN_INCOMING_                      : FALSE
• _R_CHECK_FORCE_SUGGESTS_                     : FALSE
• _R_CHECK_PACKAGES_USED_IGNORE_UNUSED_IMPORTS_: FALSE
• NOT_CRAN                                     : true
── R CMD check ───────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────
─  using log directory 'C:/Users/emily.markowitz/AppData/Local/Temp/Rtmpy80E8v/file56a01c33567f/mbquartR.Rcheck' (505ms)
─  using R version 4.4.1 (2024-06-14 ucrt)
─  using platform: x86_64-w64-mingw32
─  R was compiled by
       gcc.exe (GCC) 13.2.0
       GNU Fortran (GCC) 13.2.0
─  running under: Windows 10 x64 (build 19045)
─  using session charset: UTF-8
─  using options '--no-manual --as-cran'
✔  checking for file 'mbquartR/DESCRIPTION'
─  this is package 'mbquartR' version '0.0.0.9000'
─  package encoding: UTF-8
✔  checking package namespace information
✔  checking package dependencies (1.6s)
✔  checking if this is a source package ...
✔  checking if there is a namespace
✔  checking for executable files (1.3s)
✔  checking for hidden files and directories ...
✔  checking for portable file names
✔  checking serialization versions
✔  checking whether package 'mbquartR' can be installed (6.2s)
✔  checking installed package size (461ms)
✔  checking package directory (570ms)
✔  checking for future file timestamps ...
✔  checking 'build' directory ...
✔  checking DESCRIPTION meta-information (730ms)
✔  checking top-level files
✔  checking for left-over files ...
✔  checking index information (458ms)
✔  checking package subdirectories (1.1s)
✔  checking code files for non-ASCII characters ...
✔  checking R files for syntax errors ...
✔  checking whether the package can be loaded (668ms)
✔  checking whether the package can be loaded with stated dependencies (440ms)
✔  checking whether the package can be unloaded cleanly (549ms)
✔  checking whether the namespace can be loaded with stated dependencies (449ms)
✔  checking whether the namespace can be unloaded cleanly (651ms)
✔  checking loading without being on the library search path (1s)
✔  checking whether startup messages can be suppressed (1.2s)
✔  checking dependencies in R code (7.3s)
✔  checking S3 generic/method consistency (661ms)
✔  checking replacement functions (549ms)
✔  checking foreign function calls (533ms)
✔  checking R code for possible problems (5.2s)
✔  checking Rd files (534ms)
✔  checking Rd metadata ...
✔  checking Rd line widths ...
✔  checking Rd cross-references (439ms)
✔  checking for missing documentation entries (551ms)
✔  checking for code/documentation mismatches (1.5s)
✔  checking Rd \usage sections (998ms)
✔  checking Rd contents ...
✔  checking for unstated dependencies in examples (339ms)
✔  checking contents of 'data' directory ...
✔  checking data for non-ASCII characters (406ms)
✔  checking LazyData
✔  checking data for ASCII and uncompressed saves (395ms)
✔  checking installed files from 'inst/doc' (539ms)
✔  checking files in 'vignettes' (599ms)
─  checking examples ... [36s] OK (36.6s)
   Examples with CPU (user + system) or elapsed time > 5s
                 user system elapsed
   map_quarter  32.34   2.39   24.92
   search_coord 11.40   0.56    6.75
✔  checking for unstated dependencies in 'tests' ...
─  checking tests ...
   Running the tests in 'tests/testthat.R' failed.
     Error in `waldo::compare(x, y, ..., x_arg = x_arg, y_arg = y_arg)`: `tolerance` must be a number or `NULL`, not `TRUE`.
         ▆
      1. └─testthat::expect_equal(...) at test-search_legal.R:6:3
      2.   └─testthat:::expect_waldo_equal("equal", act, exp, info, ..., tolerance = tolerance)
      3.     └─testthat:::waldo_compare(...)
      4.       └─waldo::compare(tolerance = NULL)
      5.         └─waldo:::check_number_decimal(tolerance, allow_null = TRUE, min = 0)
      6.           └─waldo:::.stop_not_number(...)
      7.             └─waldo:::stop_input_type(...)
      8.               └─rlang::abort(message, ..., call = call, arg = arg)

✔  checking for unstated dependencies in vignettes (777ms)
✔  checking package vignettes ...
─  checking re-building of vignette outputs ... [16s] OK (16.1s)
✔  checking for non-standard things in the check directory
✔  checking for detritus in the temp directory

   See
     'C:/Users/emily.markowitz/AppData/Local/Temp/Rtmpy80E8v/file56a01c33567f/mbquartR.Rcheck/00check.log'

── R CMD check results ────────────────────────────────────────────────────────────────────────────────────────────────── mbquartR 0.0.0.9000 ────
Duration: 2m 50.7s

❯ checking tests ...
  See below...

── Test failures ─────────────────────────────────────────────────────────────────────────────────────────────────────────────────── testthat ────

> # This file is part of the standard setup for testthat.
> # It is recommended that you do not modify it.
> #
> # Where should you do additional test configuration?
> # Learn more about the roles of various files in:
> # * https://r-pkgs.org/testing-design.html#sec-tests-files-overview
> # * https://testthat.r-lib.org/articles/special-files.html
>
> library(testthat)
> library(mbquartR)
>
> test_check("mbquartR")

══ Warnings ════════════════════════════════════════════════════════════════════
── Warning ('test-mapping.R:3:3'): map_quarter() returns a map ─────────────────
package 'mapview' was built under R version 4.4.2
    ▆
 1. └─base::library(mapview) at test-mapping.R:3:3
 2.   └─base (local) testRversion(pkgInfo, package, pkgpath)

── Error ('test-search_coord.R:6:3'): search_coord() returns dataframe ─────────
Error in `waldo::compare(x, y, ..., x_arg = x_arg, y_arg = y_arg)`: `tolerance` must be a number or `NULL`, not `TRUE`.
    ▆
 1. └─testthat::expect_equal(...) at test-search_coord.R:6:3
 2.   └─testthat:::expect_waldo_equal("equal", act, exp, info, ..., tolerance = tolerance)
 3.     └─testthat:::waldo_compare(...)
 4.       └─waldo::compare(tolerance = NULL)
 5.         └─waldo:::check_number_decimal(tolerance, allow_null = TRUE, min = 0)
 6.           └─waldo:::.stop_not_number(...)
 7.             └─waldo:::stop_input_type(...)
 8.               └─rlang::abort(message, ..., call = call, arg = arg)
── Error ('test-search_legal.R:6:3'): search_legal() returns dataframe ─────────
Error in `waldo::compare(x, y, ..., x_arg = x_arg, y_arg = y_arg)`: `tolerance` must be a number or `NULL`, not `TRUE`.
    ▆
 1. └─testthat::expect_equal(...) at test-search_legal.R:6:3
 2.   └─testthat:::expect_waldo_equal("equal", act, exp, info, ..., tolerance = tolerance)
 3.     └─testthat:::waldo_compare(...)
 4.       └─waldo::compare(tolerance = NULL)
 5.         └─waldo:::check_number_decimal(tolerance, allow_null = TRUE, min = 0)
 6.           └─waldo:::.stop_not_number(...)
 7.             └─waldo:::stop_input_type(...)
 8.               └─rlang::abort(message, ..., call = call, arg = arg)

1 error ✖ | 0 warnings ✔ | 0 notes ✔
```

<summary>Test checks: 2 errors</summary>

`devtools::test()`

```
ℹ Testing mbquartR
✔ | F W  S  OK | Context
✔ |         22 | download [30.9s]
⠏ |          0 | mapping                                                                                                                          GDAL version >= 3.1.0 | setting mapviewOptions(fgb = TRUE)
✔ |   1      5 | mapping [10.2s]
──────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────
Warning (test-mapping.R:3:3): map_quarter() returns a map
package 'mapview' was built under R version 4.4.2
    ▆
 1. └─base::library(mapview) at test-mapping.R:3:3
 2.   └─base (local) testRversion(pkgInfo, package, pkgpath)
──────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────
✖ | 1        8 | search_coord [7.0s]
──────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────
Error (test-search_coord.R:6:3): search_coord() returns dataframe
Error in `waldo::compare(x, y, ..., x_arg = x_arg, y_arg = y_arg)`: `tolerance` must be a number or `NULL`, not `TRUE`.
    ▆
 1. └─testthat::expect_equal(...) at test-search_coord.R:6:3
[…]
