### 2024-12-10 #658 @alex-koiter
@jhollist I am also interested in submitting to JOSS. I know there is an opportunity to combine the rOpenSci review with the JOSS review process, is there anything in particular I should be aware of moving forward. I am currently preparing the paper.
