### 2024-12-10 #658 @jhollist
@alex-koiter I am not sure about the details on JOSS submission.  let me dig and get back to you.
