### 2024-12-10 #658 @jhollist
@alex-koiter A bit here on it: https://devguide.ropensci.org/softwarereview_author.html#preparing-for-submission.

I think we keep on with the track we are going on.  The JOSS side will happen after rOpenSci review.  The current EiC may coordinate this process.  I will find out.

In short, JOSS will make their own determination on scope and if it is in scope you will submit your paper.md for review with JOSS.  The package will not need a separate review.
