### 2024-12-10 #658 @jhollist
@alex-koiter It looks like all you will need to do is specify that the package been reviewed by rOpenSci during the pre-submisison phase with JOSS.  Here is an example issue: https://github.com/openjournals/joss-reviews/issues/6163#issuecomment-1888111491
