### 2025-02-02 #658 @maurolepore
Dear @alex-koiter and @jhollist this is just to mark the start of my EiC rotation and record what I see in case there is anything I can help with.

I see the author is still addressing reviewers' comments and the communication is fluid. I'll step back. Thanks!
