### 2025-02-26 #658 @EmilyMarkowitz-NOAA
❇ I am happy with the changes, plus these comments:

13. Sounds good about the data source being listed in the `quarters_dl` function. If interested, it might be good to have a short list in the readme. If I am pulling data from another source I always try to put that front and center. Is there a link to the data page that could be included? Maybe something as simple as:

>  Data Sources
> - The full list of quarter sections and coordinates are downloaded from the Manitoba Geoportal. Note that when the Geoportal is unavailable, data is downloaded from an archived version from the package repository on GitHub.

17. Weird that having leaflets in the reference pages is giving you trouble. In my quick search, I couldn't find an example where a leaflet output was used in the pkgdown 'Reference' section. I'll have to keep searching and will let you know if I find any examples to work/compare with.

6. I still think a map of the extent would be nice (completely a suggestion -- you're right that it might be a bit pointless) to show the boundaries of the data available. Maybe something as simple as this (I just threw it together, so if interested, please make your own!):

```
library(ggplot2)
library(sf)
library(maps)
library(rnaturalearth)

crs_out <- "EPSG:4326"
crs_in <- "+proj=longlat +datum=WGS84 +ellps=WGS84 +towgs84=0,0,0"

world_coordinates <- rnaturalearth::ne_states(c("canada", "united states of america"))  %>%
  dplyr::mutate(fill = dplyr::case_when(
    name_sv == "Manitoba" ~ "Manitoba",
    geonunit == "Canada" ~ "Canada",
    geonunit == "United States of America" ~ "USA")) %>%
  sf::st_as_sf() %>%
  sf::st_transform(crs = crs_out)

Manitoba <- world_coordinates %>%
  dplyr::filter(fill == "Manitoba")

place_labels <- data.frame(
  type = c("country", "country", "providence"),
  lab = c("United States", "Canada", "Manitoba"),
  angle = c(0, 0, 0),
  lat = c(46.405283, 52.105838, 55.989065),
  lon = c(-99.180802, -86.535079, -96.966046)) %>%
  sf::st_as_sf(coords = c("lon", "lat"),
               crs = crs_in) %>%
  sf::st_transform(crs = crs_out)

pp <- ggplot2::ggplot() +
  # plot shapes
  ggplot2::geom_sf(
    data = world_coordinates,
    mapping = aes(fill = fill),
    show.legend = FALSE,
    color = "grey20")  +
  ggplot2::scale_fill_grey() +
  # add labels
  ggplot2::geom_sf_text(
    data = place_labels %>% dplyr::filter(type == "country"),
    mapping = aes(label = lab, angle = angle),
    color = "grey40",
    fontface = "bold",
    size = 5,
    show.legend = FALSE) +
  ggplot2::geom_sf_text(
    data = place_labels %>% dplyr::filter(type == "providence"),
    mapping = aes(label = lab, angle = angle),
    color = "black",
    fontface = "bold",
    size = 6,
    show.legend = FALSE) +
  # Manage Axis extents (limits) and breaks
  ggplot2::scale_x_continuous(name = "Longitude °W",
                              breaks = seq(-120, -70, 5)) +
  ggplot2::scale_y_continuous(name = "Latitude °N",
                              breaks = seq(50, 65, 5)) + # seq(52, 62, 2)
  ggplot2::coord_sf(xlim = c(sf::st_bbox(Manitoba)[1]-5, sf::st_bbox(Manitoba)[3]+5),
                    ylim = c(sf::st_bbox(Manitoba)[2]-5, sf::st_bbox(Manitoba)[4]+5)) +
  # themes
  ggplot2::theme_bw() +
  ggplot2::theme(
    plot.margin=unit(c(0,0,0,0), "cm"),
    panel.border = element_rect(colour = "grey20", linewidth = .25, fill = NA),
    panel.background = element_rect(fill = "white"),
    panel.grid = element_line(colour="grey80", linewidth = 0.5),
    axis.text = element_text(face = "bold")
  )

pp
```
Which creates this:

![Image](https://github.com/user-attachments/assets/ae447b1c-577d-48f7-8b82-bd04e66dc89f)
