### 2025-02-27 #658 @alex-koiter
@EmilyMarkowitz-NOAA Thanks for the suggestions. I just finished making the changes.
1. Data source now listed in the README
2. Map showing the location of Manitoba is now included in the README. I misinterpreted the map suggestion, as I thought the request was to map all the quarter sections :grinning:
3. Yes I also could not find an example of where `mapview::mapview()` was used, which makes me wonder if it is expected behaviour
