### 2025-03-04 #658 @sheilasaia
👍 I'm also happy with all the changes @alex-koiter made! Yes, Emily's map extent suggestion is exactly what I was thinking of....just in case folks need a geography refresher. 😊

One last thing from me. Given Alex's and Emily's `map_quarter()` comments, I reinstalled from GitHub and when I run the code below, I'm not seeing the polygon lines in RStudio when I use `map_quarter()`. Am attaching an image of what I'm seeing and my session info. Maybe this is also related to the issue you're having for the pkgdown docs? I hope that I'm not missing something. My other thought was to ask @jhollist if he as any ideas on how to address the leaflet issue. 😄

Other than that, it looks great to me! Thank you for this opportunity to review this cool package! 🗺 💻 🧭

![Image](https://github.com/user-attachments/assets/c09ae9a3-f8da-44b5-8a1f-6327ee9bc8c6)

```
result <- search_coord(long = c(-101.4656, -99.99768, -101.4656), lat = c(51.81913, 49.928926, 51.81913))
map_quarter(result)
```

Session info:
```
> sessionInfo()
R version 4.4.2 (2024-10-31)
Platform: aarch64-apple-darwin20
Running under: macOS Sequoia 15.1.1

Matrix products: default
BLAS:   /System/Library/Frameworks/Accelerate.framework/Versions/A/Frameworks/vecLib.framework/Versions/A/libBLAS.dylib
LAPACK: /Library/Frameworks/R.framework/Versions/4.4-arm64/Resources/lib/libRlapack.dylib;  LAPACK version 3.12.0

locale:
[1] en_US.UTF-8/en_US.UTF-8/en_US.UTF-8/C/en_US.UTF-8/en_US.UTF-8

time zone: America/New_York
tzcode source: internal

attached base packages:

other attached packages:
[1] mbquartR_0.0.0.9000 gert_2.1.4          pkgreviewr_0.3.1    pkgcheck_0.1.2.077
[5] tictoc_1.2.1        fs_1.6.5            remotes_2.5.0

 [1] gtable_0.3.6            raster_3.6-31           ggplot2_3.5.1           htmlwidgets_1.6.4
 [5] lattice_0.22-6          tzdb_0.4.0              leaflet.providers_2.0.0 vctrs_0.6.5
 [9] tools_4.4.2             crosstalk_1.2.1         generics_0.1.3          stats4_4.4.2
[13] parallel_4.4.2          tibble_3.2.1            proxy_0.4-27            pkgconfig_2.0.3
[17] KernSmooth_2.23-26      satellite_1.0.5         leaflet_2.2.2           lifecycle_1.0.4
[21] farver_2.1.2            compiler_4.4.2          stringr_1.5.1           credentials_2.0.2
[25] munsell_0.5.1           terra_1.8-29            codetools_0.2-20        htmltools_0.5.8.1
[29] sys_3.4.3               class_7.3-23            jquerylib_0.1.4         pillar_1.10.1
[33] crayon_1.5.3            tidyr_1.3.1             openssl_2.3.2           classInt_0.4-11
[37] wk_0.9.4                viridis_0.6.5           tidyselect_1.2.1        digest_0.6.37
[41] stringi_1.8.4           sf_1.0-19               dplyr_1.1.4             purrr_1.0.4
[45] fastmap_1.2.0           grid_4.4.2              colorspace_2.1-1        cli_3.6.4
[49] magrittr_2.0.3          base64enc_0.1-3         utf8_1.2.4              leafem_0.2.3
[53] e1071_1.7-16            readr_2.1.5             withr_3.0.2             scales_1.3.0
[57] rappdirs_0.3.3          sp_2.2-0                bit64_4.6.0-1           bit_4.5.0.1
[61] gridExtra_2.3           png_0.1-8               askpass_1.2.1           hms_1.1.3
[65] mapview_2.11.2          viridisLite_0.4.2       s2_1.1.7                rlang_1.1.5
[69] Rcpp_1.0.14             glue_1.8.0              DBI_1.2.3               jsonlite_1.9.0
[73] rstudioapi_0.17.1       vroom_1.6.5             R6_2.6.1                units_0.8-5
```
