### 2025-03-04 #658 @alex-koiter
@sheilasaia thanks for reviewing the package again!
It looks like the polygons are in fact plotting, but you will need to zoom in on one of the points to see them. Quarter sections are very small given the size of the province. Quarter sections are only about 0.65 km<sup>2</sup>. Let me know if you can see the polygon if you zoom in on a point.
