### 2025-03-04 #658 @ropensci-review-bot
Approved! Thanks @alex-koiter for submitting and @sheilasaia, @emilymarkowitz-noaa for your reviews! :grin:
