### 2025-03-10 #658 @alex-koiter
@sheilasaia and @EmilyMarkowitz-NOAA I would like your consent to acknowledge your help in making this package so much better in the package DESCRIPTION as reviewers. If you do consent could I please get your preferred spelling of your name.

Thanks!
