### 2025-03-11 #658 @EmilyMarkowitz-NOAA
So kind, @alex-koiter ! Thank you! Yes, here is my info:

[Emily Markowitz](https://www.fisheries.noaa.gov/contact/emily-markowitz)
Research Fisheries Biologist, [Alaska Fisheries Science Center](https://www.fisheries.noaa.gov/alaska/science-data/groundfish-assessment-program-bottom-trawl-surveys)
NOAA Fisheries | U.S. Department of Commerce

Also: https://emilyhmarkowitz.github.io/emilyhmarkowitz
