### 2025-04-05 #658 @sheilasaia
Yes, that's very nice of you, @alex-koiter! Thanks so much! My info is below:

Sheila M. Saia, PhD
Environmental Data Scientist, Tetra Tech
https://sheilasaia.rbind.io/
